split crawlers tests
